Divide-and-Conquer adaptive lasso for Cox model with least square approximation

Package:	dcalasso

Type:	Package

Version:	0.1

Date:	2018-04-09

License:	GPL (>= 2)

LazyLoad:	yes

Divide-and-conquer adaptive lasso for Cox model based on divide and conquer and least square approximation

References

the paper

Author

Yan Wang yaw719@mail.harvard.edu, Tianxi Cai tcai@hsph.harvard.edu, Chuan Hong <Chuan_Hong@hms.harvard.edu>